pyproject.toml
[project]
name = "screen-translate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "opencv-python>=4.12.0.88",
    "paddleocr>=3.1.0",
    "paddlepaddle>=3.1.0",
    "pillow>=11.3.0",
    "pyautogui>=0.9.54",
    "pydantic>=2.11.7",
    "pyqt6>=6.9.1",
    "pywin32>=311",
    "pyyaml>=6.0.2",
    "sentencepiece>=0.2.1",
    "setuptools>=80.9.0",
    "transformers>=4.56.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
